+++ src/binary.rs
@@ -120,16 +120,20 @@ pub trait Binary {
 
 /// [`Binary`] implementation for a tuple of two binaries
 ///
-/// The second [`Binary`] is considered a "patch" that is only consulted if the
-/// first one did not yield an [`Instruction`]. Errors emitted always stem from
-/// the first [`Binary`].
-impl<B: Binary, P: Binary> Binary for (B, P) {
+/// This impl allows combining [`Binary`]s as long as they agree on their error
+/// type. If the first [`Binary`] returns a "miss", the second one is consulted.
+impl<A: Binary<Error = E>, B: Binary<Error = E>, E: error::MaybeMiss> Binary for (A, B) {
     type Error = B::Error;
 
     fn get_insn(&mut self, address: u64) -> Result<Instruction, Self::Error> {
-        self.0
-            .get_insn(address)
-            .or_else(|e| self.1.get_insn(address).map_err(|_| e))
+        use error::MaybeMiss;
+
+        let res = self.0.get_insn(address);
+        if res.is_miss() {
+            self.1.get_insn(address)
+        } else {
+            res
+        }
     }
 }